int to long in udf_get_last_block

+++ udf/linux-2.2/lowlevel.c
@@ -15,7 +15,7 @@
  *		ftp://prep.ai.mit.edu/pub/gnu/GPL
  *	Each contributing author retains all rights to their own work.
  *
- *  (C) 1999-2000 Ben Fennema
+ *  (C) 1999-2001 Ben Fennema
  *
  * HISTORY
  *
@@ -331,7 +331,7 @@ is_mmc(kdev_t dev, struct inode *inode_fake)
 
 #endif
 
-unsigned int
+unsigned long
 udf_get_last_block(struct super_block *sb)
 {
 	kdev_t dev = sb->s_dev;
